Added jpcmacros and wrote an initial version of getline, which is not provided on MacOSX.

+++ ahf2db.py
@@ -48,7 +48,7 @@ CREATE TABLE IF NOT EXISTS stat (
     snap_id     INTEGER,
     gid         INTEGER,
     npart       INTEGER,
-    nvpart      INTEGER,
+    nvpart      REAL,
     Xc          REAL,
     Yc          REAL,
     Zc          REAL,
